LinuxGrandfa-BE/requirements.txt
# FastAPI and ASGI server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Google Gemini AI
google-generativeai>=0.8.0

# Settings and environment management
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Form data and file uploads
python-multipart>=0.0.6

# Development dependencies
python-dotenv>=1.0.0
